--- tests/integration/auth/keeper/module_test.go
@@ -1,30 +1,0 @@
-package keeper_test
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	"cosmossdk.io/depinject"
-	"cosmossdk.io/log"
-
-	authTest "github.com/cosmos/cosmos-sdk/tests/integration/auth/keeper"
-	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
-	"github.com/cosmos/cosmos-sdk/x/auth/keeper"
-	"github.com/cosmos/cosmos-sdk/x/auth/types"
-)
-
-func TestItCreatesModuleAccountOnInitBlock(t *testing.T) {
-	var accountKeeper keeper.AccountKeeper
-	app, err := simtestutil.SetupAtGenesis(
-		depinject.Configs(
-			authTest.AppConfig,
-			depinject.Supply(log.NewNopLogger()),
-		),
-		&accountKeeper)
-	require.NoError(t, err)
-
-	ctx := app.BaseApp.NewContext(false)
-	acc := accountKeeper.GetAccount(ctx, types.NewModuleAddress(types.FeeCollectorName))
-	require.NotNil(t, acc)
-}
