--- testutil/integration/doc.go
@@ -1,3 +1,0 @@
-// Integration contains the integration test setup used for SDK modules.
-// To see how to use this, check the tests/integration/example_test.go file.
-package integration
