+++ x/bank/keeper/view.go
@@ -144,6 +144,7 @@ func (k BaseViewKeeper) GetAccountsBalances(ctx context.Context) []types.Balance
 func (k BaseViewKeeper) GetBalance(ctx context.Context, addr sdk.AccAddress, denom string) sdk.Coin {
 	amt, err := k.Balances.Get(ctx, collections.Join(addr, denom))
 	if err != nil {
+		fmt.Println(ctx)
 		return sdk.NewCoin(denom, math.ZeroInt())
 	}
 	return sdk.NewCoin(denom, amt)
